Fixed profile spin not disclosing and reward points are showing zero

+++ SampleApps/MyNTORewards/MyNTORewards/Models/ProfileModel+CurrencyPoints.swift
@@ -10,7 +10,7 @@ import LoyaltyMobileSDK
 
 extension ProfileModel {
     
-    func getCurrencyPoints(currencyName: String) -> Int {
+    func getCurrencyPoints(currencyName: String) -> Double {
         for currency in memberCurrencies {
             if currency.loyaltyMemberCurrencyName.capitalized == currencyName.capitalized {
                 return currency.pointsBalance
